Create `maker` and `taker` modules, split `order.ts` file
